+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Versions
 
+Version 2.0.2 : 68/78 Update - March 9, 2021
+
+- Move options on 78 to addons page 
+- Fix show/hide headers 
+- Fix short date format #28
+- Fix line truncation #24
+- Fix long attachment truncation #23
+- Fix unpopulated printer list on install
+
 Version 2.0.0 : 68/78 Update - January 3, 2021
 
 - Upgraded for both 68/78 WL API